Fixes in AdvancementManager.java
Removes an extra instance of When the Squad Hops into Town being listed as a hard Advancement.
Also changes the list of unreasonable Advancements to replace the duplicate How Did We Get Here? with Adventuring Time, which was missing.

I'm not entirely sure what this does but I'm pretty sure it's wrong as it is.